Upgrade dbus package to version 1.0.5
The dbus package got a new stable release with some minor changes.

This patch updates the package and fixes some minor issues that appeared
after the upgrade like `no such interface` error when requesting the '/'
object path from `org.freedesktop.DBus`. Besides that also fix the usage
of the ListNames function that would cause and endless loop until the
timeout steps in. By using the callback schema, this is prevented and
the `['finish']` way of doing things becomes obsolete.